# swagger.yaml
openapi: 3.0.0
info:
  title: HederaAgentsHub API
  description: Agent-first Digital Economy API using ERC-8004 on Hedera testnet
  version: 1.0.0
  contact:
    name: Nasrulhaq Khan
    email: your-email@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-app.up.railway.app
    description: Production server (Railway)
...
paths:
...
  /api/agents/create:
    post:
      summary: Create a new agent on Hedera blockchain
      tags:
        - Hedera Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HederaAgentCreation'
      responses:
        '200':
          description: Agent created successfully on blockchain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HederaAgentResponse'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/agents/{agentId}:
    get:
      summary: Get agent by ID from Hedera blockchain
      tags:
        - Hedera Agents
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HederaAgentResponse'
        '404':
          description: Agent not found
  /api/agents/wallet/{walletAddress}:
    get:
      summary: Get agent by wallet address (ONE AGENT PER WALLET)
      tags:
        - Hedera Agents
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
          description: Hedera wallet address
      responses:
        '200':
          description: Agent for wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HederaAgentResponse'
        '404':
          description: No agent found for wallet
  /api/agents/search/{capability}:
    get:
      summary: Search agents by capability on Hedera blockchain
      tags:
        - Hedera Agents
      parameters:
        - name: capability
          in: path
          required: true
          schema:
            type: string
          description: Capability to search for
      responses:
        '200':
          description: Agents with capability
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/HederaAgent'
                  capability:
                    type: string
                  total:
                    type: integer
  /api/agent/register:
    post:
      summary: Register a new agent
      tags:
        - Identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistration'
      responses:
        '200':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  message:
                    type: string
                    example: Agent registered successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agent/{id}:
    get:
      summary: Get agent metadata
      tags:
        - Identity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Agent metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  metadata:
                    $ref: '#/components/schemas/AgentMetadata'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agent/{id}/owner:
    get:
      summary: Get agent owner address
      tags:
        - Identity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Agent owner
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  owner:
                    type: string
                    example: '0x98d46dda75A216f94bF9784a6665A4E3821Da3D9'
        '404':
          description: Agent not found
  /api/agents/search:
    get:
      summary: Search agents by capability
      tags:
        - Identity
      parameters:
        - name: capability
          in: query
          required: true
          schema:
            type: string
          description: Capability to search for
          example: weather
      responses:
        '200':
          description: List of matching agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  capability:
                    type: string
                    example: weather
                  agents:
                    type: array
                    items:
                      type: string
                    example: ['agent_1', 'agent_3']
                  count:
                    type: integer
                    example: 2
        '400':
          description: Missing capability parameter
  /api/agents/count/{address}:
    get:
      summary: Get agent count for address
      tags:
        - Identity
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Owner address
          example: '0x98d46dda75A216f94bF9784a6665A4E3821Da3D9'
      responses:
        '200':
          description: Agent count
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  address:
                    type: string
                    example: '0x98d46dda75A216f94bF9784a6665A4E3821Da3D9'
                  count:
                    type: integer
                    example: 3
  /api/feedback:
    post:
      summary: Submit feedback for an agent
      tags:
        - Reputation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackSubmission'
      responses:
        '200':
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  feedbackId:
                    type: string
                    example: feedback_1
                  message:
                    type: string
                    example: Feedback submitted successfully
        '400':
          description: Bad request - invalid rating or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reputation/{agentId}:
    get:
      summary: Get reputation summary
      tags:
        - Reputation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Reputation summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  reputation:
                    $ref: '#/components/schemas/ReputationScore'
        '404':
          description: No reputation data found
  /api/reputation/{agentId}/trust:
    get:
      summary: Get trust score and trustworthiness
      tags:
        - Reputation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Trust score
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  trustScore:
                    type: integer
                    example: 93
                    description: Trust score (0-100)
                  trustworthy:
                    type: boolean
                    example: true
  /api/feedback/{agentId}:
    get:
      summary: Get all feedback IDs for an agent
      tags:
        - Reputation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: List of feedback IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  feedbackIds:
                    type: array
                    items:
                      type: string
                    example: ['feedback_1', 'feedback_2', 'feedback_3']
                  count:
                    type: integer
                    example: 3
  /api/validation/request:
    post:
      summary: Request validation for an agent
      tags:
        - Validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation requested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  validationId:
                    type: string
                    example: validation_1
                  message:
                    type: string
                    example: Validation requested successfully
        '400':
          description: Missing required fields
  /api/validation/submit:
    post:
      summary: Submit validation result
      tags:
        - Validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationSubmission'
      responses:
        '200':
          description: Validation result submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  validationId:
                    type: string
                    example: validation_1
                  message:
                    type: string
                    example: Validation result submitted successfully
        '400':
          description: Missing required fields
  /api/validation/{agentId}/score:
    get:
      summary: Get validation score
      tags:
        - Validation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Validation score
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  validationScore:
                    $ref: '#/components/schemas/ValidationScore'
        '404':
          description: No validation data found
  /api/validation/{agentId}/confidence:
    get:
      summary: Get validation confidence
      tags:
        - Validation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent ID
          example: agent_1
      responses:
        '200':
          description: Validation confidence
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  agentId:
                    type: string
                    example: agent_1
                  confidence:
                    type: integer
                    example: 100
                    description: Confidence score (0-100)
                  validated:
                    type: boolean
                    example: true
components:
  schemas:
    HederaAgentCreation:
      type: object
      required:
        - name
        - purpose
        - capabilities
        - walletAddress
      properties:
        name:
          type: string
          example: iPhone Shopping Assistant
          description: Agent name
        purpose:
          type: string
          example: You are an intelligent shopping assistant...
          description: Agent purpose (system prompt)
        capabilities:
          type: array
          items:
            type: string
          example: ['product-search', 'price-comparison', 'payment-processing']
          description: Agent capabilities
        walletAddress:
          type: string
          example: '0.0.7174687'
          description: Hedera wallet address (ONE AGENT PER WALLET)
        metadata:
          type: object
          example:
            version: '1.0.0'
            model: 'gpt-4'
            temperature: 0.7
          description: Optional metadata
    HederaAgentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        agentId:
          type: string
          example: agent-iphone-shopping-assistant-1761990620227-hqa6d
        name:
          type: string
          example: iPhone Shopping Assistant
        purpose:
          type: string
          example: You are an intelligent shopping assistant...
        capabilities:
          type: array
          items:
            type: string
          example: ['product-search', 'price-comparison']
        walletAddress:
          type: string
          example: '0.0.7174687'
        evmAddress:
          type: string
          example: '0x98d46dda75A216f94bF9784a6665A4E3821Da3D9'
          description: EVM address for wallet access
        privateKey:
          type: string
          example: '0xab9c734fb98648d32634d96e9a1629fe06f676eb7b940b6e...'
          description: Private key for wallet operations
        topicId:
          type: string
          example: '0.0.7178279'
          description: Hedera Topic ID where agent is stored
        transactionId:
          type: string
          example: '0.0.7174687@1761990615.794170000'
          description: Blockchain transaction ID
        error:
          type: string
          description: Error message if creation failed
    HederaAgent:
      type: object
      properties:
        agentId:
          type: string
          example: agent-iphone-shopping-assistant-1761990620227-hqa6d
        name:
          type: string
          example: iPhone Shopping Assistant
        purpose:
          type: string
          example: You are an intelligent shopping assistant...
        capabilities:
          type: array
          items:
            type: string
          example: ['product-search', 'price-comparison']
        walletAddress:
          type: string
          example: '0.0.7174687'
        topicId:
          type: string
          example: '0.0.7178279'
        transactionId:
          type: string
          example: '0.0.7174687@1761990615.794170000'
        createdAt:
          type: number
          example: 1761990620227
    AgentRegistration:
      type: object
      required:
        - name
        - description
        - capabilities
      properties:
        name:
          type: string
          example: Weather Bot
          description: Agent name
        description:
          type: string
          example: Provides real-time weather data
          description: Agent description
        capabilities:
          type: array
          items:
            type: string
          example: ['weather', 'forecast', 'alerts']
          description: Agent capabilities
    AgentMetadata:
      type: object
      properties:
        name:
          type: string
          example: Weather Bot
        description:
          type: string
          example: Provides real-time weather data
        capabilities:
          type: array
          items:
            type: string
          example: ['weather', 'forecast', 'alerts']
    FeedbackSubmission:
      type: object
      required:
        - agentId
        - rating
        - comment
      properties:
        agentId:
          type: string
          example: agent_1
          description: Agent ID to rate
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
          description: Rating (1-5 stars)
        comment:
          type: string
          example: Excellent service!
          description: Feedback comment
        paymentProof:
          type: string
          example: '0x402_proof_hash'
          description: Optional x402 payment proof
    ReputationScore:
      type: object
      properties:
        agentId:
          type: string
          example: agent_1
        averageRating:
          type: number
          format: float
          example: 4.67
          description: Average rating (0-5)
        totalReviews:
          type: integer
          example: 3
          description: Total number of reviews
        trustScore:
          type: integer
          example: 93
          description: Trust score (0-100)
    ValidationRequest:
      type: object
      required:
        - agentId
        - validationType
        - description
      properties:
        agentId:
          type: string
          example: agent_1
          description: Agent ID to validate
        validationType:
          type: string
          enum:
            - stake-re-execution
            - zkml-proof
            - tee-oracle
            - trusted-judge
            - multi-sig
          example: stake-re-execution
          description: Type of validation method
        description:
          type: string
          example: Verify weather data accuracy
          description: Validation description
        stake:
          type: integer
          format: int64
          example: 1000000000000000000
          description: Stake amount in wei (optional)
    ValidationSubmission:
      type: object
      required:
        - validationId
        - isValid
        - evidence
      properties:
        validationId:
          type: string
          example: validation_1
          description: Validation ID
        isValid:
          type: boolean
          example: true
          description: Validation result
        evidence:
          type: string
          example: Verified - data matches expected output
          description: Validation evidence
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
          description: Error description